Update state diagram method.
